Project Euler #4 Largest Palindrome Product
This update adds the next programming challenge that I completed. The purpose of this project, was to take two 3 digit numbers, multiply them together and determine if the number was a palindrome or not. If it was it needs to make sure it is the largest palindrome and then print the answer to console.