Largest Product in a Grid
Find the four adjacent numbers in the same direction that produce the largest sum.  This program, loops through the grid and does equations to find which four adjacent numbers produce the largest sum. Then it saves it and moves on, if it finds a sum that is larger, its saves it until either the end of the loops or it finds a new largest sum.